requirements.txt
Django==2.2.1
django-location-field==2.1.0
psycopg2==2.8.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
